Small cleanups and a test suite
Add a test suite for the current routes, which requires mocking
Butler and the Google Cloud Storage API.  Return 404 if asked for
links for an object that isn't found, rather than returning a
standard links structure with a null URL.  Flesh out the OpenAPI
documentation a bit and make some coding style tweaks.

Enable verbosity for pytest in tox so that it will show output
diffs.