Add MIN and MAX for CIRCLE cutout parameter
We want to limit the radius to 0.25 degrees for better display in
the Portal (this is slightly larger than a CCD/patch-size image).
Specifying limits for CIRCLE requires specifying all three values,
so limit ra to -180 to 180 and dec to -90 to 90.  (Ideally we
would do better based on the coordinates of the image in question,
but the data for this is not yet available.)