Don't return cutout service_ref for raws
Raw images aren't supported by the cutout service.  Detect those
with the Butler DatasetRef and exclude the cutout service_ref from
the returned links table.

This is a stopgap solution for DP0.2 until we design a more modular
architecture and decide how much information we want to encode in
the obstap table.