Add support for generating the HiPS list
The HiPS list accumulates the properties files for all of the HiPS
data sets for a given deployment of the Rubin Science Platform,
with the addition of the URL (which varies by environment and thus
cannot be encoded in the properties file of the individual HiPS
data set).  Serve that list by getting the individual properties
files of each data set, assembling, and then caching the list.

Including this in datalinker may look like an odd choice, since it's
unrelated to the DataLink standard.  But, viewed more broadly, the
function of datalinker is to provide registries, catalogs, and
redirects to other services in the deployment, and HiPS is another
example of that type of service.  The HiPS data sets themselves will
be served directly from GCS in the future, and GCS won't be able to
serve the dynamic HiPS list directly, so we need some small web
application to do so.  Including this in datalinker seems better
than writing a separate application solely for that purpose.