[DM-39905] Get datalinker build working
First, safir now requires python 3.11, so
to get to any version above 3.8.0, you need
to change the tox.ini to force it to require
python 3.11.  Previously tox.ini was using
the system python and not forcing to python3.11
even if it was installed.  Then safir would
fail to install.

Also update the dependencies in the requirements
directory.