[DM-40192] Explicitly pass in S3 endpoint
Apparently with boto the automatic and using environment variables
for the S3_ENDPOINT_URL only works and configures itself if you
are using aws.  Since for our case we're using our own S3, we
need to pass it in when constructing the client.

Also, fix an unneeded dependency on the test_links_errors_s3 test.