Test with real ResourcePath instead of a mock
In the next commit I will be adding a usage of another ResourcePath property, so instead of mocking it there is no reason not to use the real library.

There was also a convoluted way of specifying the test size that was effectively a constant, so just make it a constant.