Allow overwrite of tap metadata
It seems to be possible for a datalinker container to get in a state where it is trying to start up, but fails because it is trying to download TAP metadata for a second time and unzip interactively prompts to overwrite.  Add the allow-overwrite flag so it can get past this state.