Convert to Ruff and fix all issues
Convert to Ruff from Black, isort, and flake8. Fix all of the
issues that it complained about. Use Annotated for all arguments
to FastAPI handler functions, which is now recommended.

Convert the configuration class to use pydantic-settings and its
support for environment variables. Use an enum for the storage
backend, which adds validation and removes the need to handle an
unknown storage backend error explicitly.

Convert fixtures that change environment variables or configuration
settings to use pytest's monkeypatch fixture.