Tighten configuration settings and validation
Use a DATALINKER_ prefix for all configuration environment variables
and stop using SAFIR_ as a prefix for settings that are passed to
Safir. Make mandatory environment variables actually mandatory, and
set them to dummy values in tox to avoid issues with the test suite.
Declare URLs as HttpUrl so that Pydantic will validate them.